backend/requirements.txt
# ==================== Core Framework ====================
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-dotenv==1.0.0

# ==================== Database ====================
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# ==================== Data Validation ====================
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# ==================== Data Processing ====================
numpy==1.26.3
pandas==2.1.4
openpyxl==3.1.2
chardet==5.2.0
python-dateutil==2.8.2

# ==================== Export/Reporting ====================
reportlab==4.0.7
matplotlib==3.8.2
Pillow==10.2.0

# ==================== NLP ====================
spacy==3.7.2
fr-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/fr_core_news_sm-3.7.0/fr_core_news_sm-3.7.0-py3-none-any.whl
nltk==3.8.1
scikit-learn==1.3.2

# ==================== Document Processing ====================
PyPDF2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
markdown==3.5.1

# ==================== HTTP & Async ====================
httpx
aiohttp==3.9.1
requests==2.31.0

# ==================== Utilities ====================
tenacity==8.2.3
pytz==2023.3

# ==================== Development / Testing ====================
pytest==7.4.3
pytest-asyncio==0.21.1

# ==================== RAG System Dependencies ====================
# LlamaIndex Core
llama-index==0.10.12
llama-index-core==0.10.12

# LlamaIndex Integrations
llama-index-llms-ollama==0.1.2
llama-index-embeddings-ollama==0.1.2
llama-index-vector-stores-faiss==0.1.2
llama-index-readers-file==0.1.4

# Vector Store
faiss-cpu==1.7.4

# Redis Cache
redis==5.0.1

# Additional dependencies for document processing
tiktoken==0.5.2

# =================== OCR ====================
loguru==0.7.2
pytesseract
ollama
opencv-python==4.8.1.78
pdf2image==1.17.0

# =================== Machine Learning ====================

scikit-learn==1.3.2
xgboost==2.0.3

prophet==1.1.5
statsmodels==0.14.1

optuna==3.5.0  # For hyperparameter tuning
joblib==1.3.2  # For model persistence

seaborn==0.13.1
plotly==5.18.0

scipy==1.11.4

# ==================== Auth ====================
python-jose[cryptography]==3.3.0

# ==================== Other ====================
groq>=0.4.0
networkx>=3.2
PyMuPDF>=1.23.21

